// bookspace-client/src/types/review-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import type {BookBarebonesDTO} from './book-barebones-dto';
 /**
 * 
 *
 * @export
 * @interface ReviewDTO
 */
export interface ReviewDTO {

    /**
     * @type {number}
     * @memberof ReviewDTO
     */
    id?: number;

    /**
     * @type {BookBarebonesDTO}
     * @memberof ReviewDTO
     */
    book?: BookBarebonesDTO;

    /**
     * @type {string}
     * @memberof ReviewDTO
     */
    username?: string;

    /**
     * @type {string}
     * @memberof ReviewDTO
     */
    comment?: string;

    /**
     * @type {number}
     * @memberof ReviewDTO
     */
    rating?: number;

    /**
     * @type {Date}
     * @memberof ReviewDTO
     */
    createdAt?: Date;
}